Add authentication blueprint and models
This commit adds a new authentication blueprint and models for the web application. It includes routes for login, signup, and logout, as well as models for Music and WebDAV.
