+++ web/auth.py
@@ -1,0 +1,16 @@
+from flask import Blueprint
+from . import db
+
+auth = Blueprint('auth', __name__)
+
+@auth.route('/login')
+def login():
+    return 'Login'
+
+@auth.route('/signup')
+def signup():
+    return 'Signup'
+
+@auth.route('/logout')
+def logout():
+    return 'Logout'
