+++ web/models.py
@@ -1,0 +1,15 @@
+from . import db
+
+class Music(db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    title = db.Column(db.String(100))
+    url = db.Column(db.String(200))
+    monitored = db.Column(db.Boolean)
+    interval = db.Column(db.Integer)
+
+class WebDAV(db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    WebDAV_URL = db.Column(db.String(250))
+    WebDAV_Directory = db.Column(db.String(250))
+    WebDAV_Username = db.Column(db.String(30))
+    WebDAV_Password = db.Column(db.String(100))
